pyproject.toml
[project]
name = "trimcloud"
version = "0.1.0"
description = ""
authors = [
    {name = "Nsuidara",email = "nsuidara@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
